A very simple status display facility
A status manager maintains a queue of status items. Other facitlities
can push and pop status items to indicate a status change. Status items
can also be set to be popped off after a given period.

A status view controller manages the display of status items maintained
by a status manager. The default status view controller provides a HUD
notification style similar to what can be seen on iOS when changing the
volume using the volume buttons on the device.